style: unify game stage layouts and improve UI consistency
- Add consistent vertical stack layout across all game stages
- Implement matching button styles with wiggle animation
- Create uniform loading states and image containers
- Style prompt options in GuessingStage as single column
- Remove inline styles in favor of CSS classes
- Add hover and active states for interactive elements
- Standardize typography, spacing, and colors
